deletBlog unsnubbed

+++ src/main/java/com/driver/services/BlogService.java
@@ -74,6 +74,7 @@ public class BlogService {
 
     public void deleteBlog(int blogId){
         //delete blog and corresponding images
+        if(blogRepository1.findById(blogId)==null) return;
         blogRepository1.deleteById(blogId);
     }
 }
